Update dashboard navigation and add new pages
The navigation links in the dashboard have been updated to use the 'next/link' instead of 'a' tag. This includes the addition of conditional styling based on the current path. New pages for Invoices, Customers, and the Dashboard itself have been added. Layout and main page also received some enhancements and added logos and images.